Cleaner code.
